refactor: DBTI 결과 UI 모델 분리 및 매퍼 추가
- `ResultBox.kt` 내부에 정의되어 있던 `TraitAnalysis` 데이터 클래스를 별도 파일(`TraitAnalysis.kt`)로 분리
- `DbtiAnalysisEntity`를 `TraitAnalysis` UI 모델로 변환하는 매퍼 함수(`toUiModel`) 추가
