+++ app/src/main/java/com/paw/key/presentation/ui/dbti/result/model/TraitAnalysis.kt
@@ -1,0 +1,18 @@
+package com.paw.key.presentation.ui.dbti.result.model
+
+import com.paw.key.domain.entity.DBTI.DbtiAnalysisEntity
+
+// TraitAnalysis 데이터 클래스 정의
+data class TraitAnalysis(
+    val leftLabel: String,
+    val rightLabel: String,
+    val dominantSide: String,
+    val score: Int
+)
+
+fun DbtiAnalysisEntity.toUiModel() = TraitAnalysis(
+    leftLabel = leftLabel,
+    rightLabel = rightLabel,
+    dominantSide = dominantSide,
+    score = score
+)
